Update preferences table migration to use unsigned integer for course_id

+++ backend/database/migrations/2024_08_07_005134_create_preferences_table.php
@@ -10,13 +10,13 @@ return new class extends Migration {
         Schema::create('preferences', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('faculty_id');
-            $table->unsignedBigInteger('course_id');
+            $table->unsignedInteger('course_id'); 
             $table->string('preferred_day');
             $table->string('preferred_time');
             $table->timestamps();
 
             $table->foreign('faculty_id')->references('id')->on('faculty')->onDelete('cascade');
-            $table->foreign('course_id')->references('course_id')->on('courses')->onDelete('cascade'); // Adjusted to match Course model
+            $table->foreign('course_id')->references('course_id')->on('courses')->onDelete('cascade');
         });
     }
 
@@ -25,4 +25,3 @@ return new class extends Migration {
         Schema::dropIfExists('preferences');
     }
 };
-
